# openapi.yaml
openapi: 3.1.0
info:
  title: assessment-service
  version: v0.3.1
paths:
...
  /v1/admin/tests:
    get:
      tags:
      - assessment
      - assessment
      summary: List Tests
      operationId: getAdminTests
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TestResponse'
                type: array
                title: Response List Tests V1 Admin Tests Get
              example:
              - test_id: 00000000-0000-4000-8000-000000000000
                subject_code: string
                grade: 1
                version: 1
                questions:
                - {}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
    post:
      tags:
      - assessment
      - assessment
      summary: Create Test
      operationId: postAdminTests
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestRequest'
            example:
              subject_code: string
              grade: 1
              questions:
              - node_id: string
                text: string
                answer_key: string
                correct_option_id: string
                options: []
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResponse'
              example:
                test_id: 00000000-0000-4000-8000-000000000000
                subject_code: string
                grade: 1
                version: 1
                questions:
                - question_id: 00000000-0000-4000-8000-000000000000
                  node_id: string
                  text: string
                  max_score: 1
                  options: []
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
  /v1/tests:
    get:
      tags:
      - assessment
      - assessment
      summary: List Tests
      operationId: getTests
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TestResponse'
                type: array
                title: Response List Tests V1 Tests Get
              example:
              - test_id: 00000000-0000-4000-8000-000000000000
                subject_code: string
                grade: 1
                version: 1
                questions:
                - {}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
    post:
      tags:
      - assessment
      - assessment
      summary: Create Test
      operationId: postTests
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestRequest'
            example:
              subject_code: string
              grade: 1
              questions:
              - node_id: string
                text: string
                answer_key: string
                correct_option_id: string
                options: []
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestResponse'
              example:
                test_id: 00000000-0000-4000-8000-000000000000
                subject_code: string
                grade: 1
                version: 1
                questions:
                - question_id: 00000000-0000-4000-8000-000000000000
                  node_id: string
                  text: string
                  max_score: 1
                  options: []
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
  /v1/admin/tests/{test_id}/publish:
    post:
      tags:
      - assessment
      - assessment
      summary: Publish Test
      operationId: postAdminTestsPublishByTestId
      parameters:
      - name: test_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Test Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishTestResponse'
              example:
                test_id: 00000000-0000-4000-8000-000000000000
                status: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    CreateTestRequest:
      properties:
        subject_code:
          type: string
          title: Subject Code
        grade:
          type: integer
          maximum: 4.0
          minimum: 1.0
          title: Grade
        questions:
          items:
            $ref: '#/components/schemas/QuestionCreateRequest'
          type: array
          title: Questions
      type: object
      required:
      - subject_code
      - grade
      - questions
      title: CreateTestRequest
...
    DiagnosticTag:
      type: string
      enum:
      - inattention
      - misread_condition
      - calc_error
      - concept_gap
      - guessing
      - other
      title: DiagnosticTag
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PublishTestResponse:
      properties:
        test_id:
          type: string
          format: uuid
          title: Test Id
        status:
          type: string
          title: Status
      type: object
      required:
      - test_id
      - status
      title: PublishTestResponse
    QuestionCreateRequest:
      properties:
        node_id:
          type: string
          title: Node Id
        text:
          type: string
          title: Text
        question_type:
          $ref: '#/components/schemas/QuestionType'
          default: text
        answer_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Answer Key
        correct_option_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Correct Option Id
        options:
          items:
            $ref: '#/components/schemas/QuestionOptionCreateRequest'
          type: array
          title: Options
        text_distractors:
          items:
            $ref: '#/components/schemas/TextDistractorCreateRequest'
          type: array
          title: Text Distractors
        max_score:
          type: integer
          title: Max Score
          default: 1
      type: object
      required:
      - node_id
      - text
      title: QuestionCreateRequest
    QuestionOptionCreateRequest:
      properties:
        option_id:
          type: string
          title: Option Id
        text:
          type: string
          title: Text
        position:
          type: integer
          minimum: 1.0
          title: Position
        diagnostic_tag:
          anyOf:
          - $ref: '#/components/schemas/DiagnosticTag'
          - type: 'null'
      type: object
      required:
      - option_id
      - text
      - position
      title: QuestionOptionCreateRequest
    QuestionOptionResponse:
      properties:
        option_id:
          type: string
          title: Option Id
        text:
          type: string
          title: Text
        position:
          type: integer
          title: Position
      type: object
      required:
      - option_id
      - text
      - position
      title: QuestionOptionResponse
    QuestionResponse:
      properties:
        question_id:
          type: string
          format: uuid
          title: Question Id
        node_id:
          type: string
          title: Node Id
        text:
          type: string
          title: Text
        question_type:
          $ref: '#/components/schemas/QuestionType'
        max_score:
          type: integer
          title: Max Score
        options:
          items:
            $ref: '#/components/schemas/QuestionOptionResponse'
          type: array
          title: Options
      type: object
      required:
      - question_id
      - node_id
      - text
      - question_type
      - max_score
      title: QuestionResponse
    QuestionType:
      type: string
      enum:
      - text
      - single_choice
      title: QuestionType
...
    TestResponse:
      properties:
        test_id:
          type: string
          format: uuid
          title: Test Id
        subject_code:
          type: string
          title: Subject Code
        grade:
          type: integer
          title: Grade
        version:
          type: integer
          title: Version
        questions:
          items:
            $ref: '#/components/schemas/QuestionResponse'
          type: array
          title: Questions
      type: object
      required:
      - test_id
      - subject_code
      - grade
      - version
      - questions
      title: TestResponse
    TextDistractorCreateRequest:
      properties:
        pattern:
          type: string
          title: Pattern
        match_mode:
          $ref: '#/components/schemas/TextMatchMode'
          default: exact
        diagnostic_tag:
          $ref: '#/components/schemas/DiagnosticTag'
          default: other
      type: object
      required:
      - pattern
      title: TextDistractorCreateRequest
    TextMatchMode:
      type: string
      enum:
      - exact
      - normalized
      - regex
      title: TextMatchMode
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ProblemDetails:
      type: object
      title: ProblemDetails
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        request_id:
          type: string
        code:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://example.com/problems/authentication
            title: Unauthorized
            status: 401
            detail: Bearer access token is required
            code: UNAUTHORIZED
            request_id: 00000000-0000-4000-8000-000000000000
    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://example.com/problems/access-denied
            title: Forbidden
            status: 403
            detail: Access denied
            code: FORBIDDEN
            request_id: 00000000-0000-4000-8000-000000000000
    TooManyRequests:
      description: Too Many Requests
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: https://example.com/problems/too-many-requests
            title: Too Many Requests
            status: 429
            detail: Rate limit exceeded
            code: RATE_LIMITED
            request_id: 00000000-0000-4000-8000-000000000000
servers:
- url: http://localhost:8003
  description: Local
- url: https://api.monitoring.example
  description: Production
security:
- bearerAuth: []